Fix D-Bus related crash in kttsd
The problem was that kttsd registered its D-Bus /KSpeech object
as soon as the application started, without making sure that the
daemon is fully initialized.

When during initialization of the daemon a KMessageBox is put up
(asking the user if he wants to add talkers to make speech work),
that message box tries to connect to the speech daemon via a D-Bus
call.

We now delay the registration of the object until we are fully
initialized. The other option was to entirely remove the message
box (which is bad, as the user may want to know what he needs to
do to get speech work), or make the message box non-modal, which
I found too complicated...

BUG: 164518

svn path=/trunk/KDE/kdeaccessibility/kttsd/; revision=972345